migrate to springboot #1849
