+++ application/wlp-web-java-spring/src/main/java/webapp/tier/WlpWebJavaSpringApplication.java
@@ -1,0 +1,13 @@
+package webapp.tier;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class WlpWebJavaSpringApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(WlpWebJavaSpringApplication.class, args);
+	}
+
+}
