+++ application/wlp-web-java-spring/src/main/java/webapp/tier/controller/AdviceController.java
@@ -1,4 +1,4 @@
-package webapp.controller;
+package webapp.tier.controller;
 
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
