+++ application/wlp-web-java-spring/src/main/java/webapp/tier/resource/RedisResource.java
@@ -1,4 +1,4 @@
-package webapp.resource;
+package webapp.tier.resource;
 
 import javax.ws.rs.Consumes;
 import javax.ws.rs.GET;
@@ -8,21 +8,21 @@ import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 
-import webapp.tier.cache.MemcachedService;
+import webapp.tier.cache.RedisService;
 
-@Path("/memcached")
+@Path("/redis")
 @Produces(MediaType.APPLICATION_JSON)
 @Consumes(MediaType.APPLICATION_JSON)
-public class MemcachedResource {
+public class RedisResource {
 
-	MemcachedService createMemcachedService() {
-		return new MemcachedService();
+	RedisService createRedisService() {
+		return new RedisService();
 	}
 
 	@POST
 	@Path("/set")
 	public Response set() {
-		MemcachedService svc = createMemcachedService();
+		RedisService svc = createRedisService();
 		try {
 			return Response.ok().entity(svc.set()).build();
 		} catch (Exception e) {
@@ -34,9 +34,21 @@ public class MemcachedResource {
 	@GET
 	@Path("/get")
 	public Response get() {
-		MemcachedService svc = createMemcachedService();
+		RedisService svc = createRedisService();
 		try {
 			return Response.ok().entity(svc.get()).build();
+		} catch (Exception e) {
+			e.printStackTrace();
+			return Response.status(500).build();
+		}
+	}
+
+	@POST
+	@Path("/publish")
+	public Response publish() {
+		RedisService svc = createRedisService();
+		try {
+			return Response.ok().entity(svc.publish()).build();
 		} catch (Exception e) {
 			e.printStackTrace();
 			return Response.status(500).build();
