+++ application/wlp-web-java-spring/src/test/java/webapp/tier/cache/RedisServiceTest.java
@@ -3,7 +3,7 @@ package webapp.tier.cache;
 import static org.hamcrest.CoreMatchers.containsString;
 import static org.hamcrest.CoreMatchers.instanceOf;
 import static org.hamcrest.CoreMatchers.is;
-import static org.junit.Assert.assertThat;
+import static org.hamcrest.MatcherAssert.assertThat;
 import static org.junit.jupiter.api.Assertions.fail;
 
 import java.io.IOException;
