+++ application/wlp-web-java-spring/src/test/java/webapp/tier/config/JaxrsConfigTest.java
@@ -1,19 +1,19 @@
-package webapp.config;
+package webapp.tier.config;
 
-import static org.hamcrest.CoreMatchers.*;
-import static org.junit.Assert.*;
+import static org.hamcrest.CoreMatchers.is;
+import static org.hamcrest.MatcherAssert.assertThat;
 
 import java.util.Set;
 
 import org.junit.jupiter.api.Test;
 
-import webapp.resource.ActiveMqResource;
-import webapp.resource.HazelcastResource;
-import webapp.resource.MemcachedResource;
-import webapp.resource.MysqlResource;
-import webapp.resource.PostgresResource;
-import webapp.resource.RabbitmqResource;
-import webapp.resource.RedisResource;
+import webapp.tier.resource.ActiveMqResource;
+import webapp.tier.resource.HazelcastResource;
+import webapp.tier.resource.MemcachedResource;
+import webapp.tier.resource.MysqlResource;
+import webapp.tier.resource.PostgresResource;
+import webapp.tier.resource.RabbitmqResource;
+import webapp.tier.resource.RedisResource;
 
 class JaxrsConfigTest {
