+++ application/wlp-web-java-spring/src/test/java/webapp/tier/controller/ApplicationControllerTest.java
@@ -1,6 +1,6 @@
-package webapp.controller;
+package webapp.tier.controller;
 
-import static org.junit.Assert.fail;
+import static org.junit.jupiter.api.Assertions.fail;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.model;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
