+++ application/wlp-web-java-spring/src/test/java/webapp/tier/db/PostgresServiceTest.java
@@ -3,7 +3,7 @@ package webapp.tier.db;
 import static org.hamcrest.CoreMatchers.containsString;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.junit.Assert.fail;
+import static org.junit.jupiter.api.Assertions.fail;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
