+++ application/wlp-web-java-spring/src/test/java/webapp/tier/mq/RabbitmqServiceTest.java
@@ -2,8 +2,8 @@ package webapp.tier.mq;
 
 import static org.hamcrest.CoreMatchers.containsString;
 import static org.hamcrest.CoreMatchers.is;
-import static org.junit.Assert.assertThat;
-import static org.junit.Assert.fail;
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.junit.jupiter.api.Assertions.fail;
 
 import java.util.Objects;
