+++ application/wlp-web-java-spring/src/test/java/webapp/tier/resource/MysqlResourceTest.java
@@ -1,9 +1,9 @@
-package webapp.resource;
+package webapp.tier.resource;
 
 import static org.hamcrest.CoreMatchers.instanceOf;
 import static org.hamcrest.CoreMatchers.is;
-import static org.junit.Assert.assertThat;
-import static org.junit.Assert.fail;
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.junit.jupiter.api.Assertions.fail;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
@@ -14,37 +14,37 @@ import javax.ws.rs.core.Response;
 
 import org.junit.jupiter.api.Test;
 
-import webapp.tier.db.PostgresService;
+import webapp.tier.db.MysqlService;
 
-public class PostgresResourceTest {
+public class MysqlResourceTest {
 
-	private PostgresResource createPostgresService() throws Exception {
-		PostgresService svc = mock(PostgresService.class);
+	private MysqlResource createMysqlService() throws Exception {
+		MysqlService svc = mock(MysqlService.class);
 		List<String> allmsg = new ArrayList<>();
 		allmsg.add("OK");
 		when(svc.select()).thenReturn(allmsg);
 		when(svc.insert()).thenReturn("OK");
 		when(svc.delete()).thenReturn("OK");
-		return new PostgresResource() {
-			PostgresService createPostgresService() {
+		return new MysqlResource() {
+			MysqlService createMysqlService() {
 				return svc;
 			}
 		};
 	}
 
-	private PostgresResource createPostgresServiceNull() throws Exception {
-		return new PostgresResource() {
-			PostgresService createPostgresService() {
+	private MysqlResource createMysqlServiceNull() throws Exception {
+		return new MysqlResource() {
+			MysqlService createMysqlService() {
 				return null;
 			}
 		};
 	}
 
 	@Test
-	public void testcreatePostgresService() {
+	public void testcreateMysqlService() {
 		try {
-			PostgresResource rsc = new PostgresResource();
-			assertThat(rsc.createPostgresService(), is(instanceOf(PostgresService.class)));
+			MysqlResource rsc = new MysqlResource();
+			assertThat(rsc.createMysqlService(), is(instanceOf(MysqlService.class)));
 		} catch (Exception e) {
 			e.printStackTrace();
 			fail();
@@ -54,7 +54,7 @@ public class PostgresResourceTest {
 	@Test
 	public void testinsert() {
 		try {
-			PostgresResource rsc = createPostgresService();
+			MysqlResource rsc = createMysqlService();
 			Response resp = rsc.insert();
 			assertThat(resp.getStatus(), is(200));
 			assertThat(resp.getEntity().toString(), is("OK"));
@@ -67,7 +67,7 @@ public class PostgresResourceTest {
 	@Test
 	public void testinsertError() {
 		try {
-			PostgresResource rsc = createPostgresServiceNull();
+			MysqlResource rsc = createMysqlServiceNull();
 			Response resp = rsc.insert();
 			assertThat(resp.getStatus(), is(500));
 		} catch (Exception e) {
@@ -79,7 +79,7 @@ public class PostgresResourceTest {
 	@Test
 	public void testselect() {
 		try {
-			PostgresResource rsc = createPostgresService();
+			MysqlResource rsc = createMysqlService();
 			Response resp = rsc.select();
 			assertThat(resp.getStatus(), is(200));
 			assertThat(resp.getEntity().toString(), is("[OK]"));
@@ -92,7 +92,7 @@ public class PostgresResourceTest {
 	@Test
 	public void testselectError() {
 		try {
-			PostgresResource rsc = createPostgresServiceNull();
+			MysqlResource rsc = createMysqlServiceNull();
 			Response resp = rsc.select();
 			assertThat(resp.getStatus(), is(500));
 		} catch (Exception e) {
@@ -104,7 +104,7 @@ public class PostgresResourceTest {
 	@Test
 	public void testdelete() {
 		try {
-			PostgresResource rsc = createPostgresService();
+			MysqlResource rsc = createMysqlService();
 			Response resp = rsc.delete();
 			assertThat(resp.getStatus(), is(200));
 			assertThat(resp.getEntity().toString(), is("OK"));
@@ -117,7 +117,7 @@ public class PostgresResourceTest {
 	@Test
 	public void testdeleteError() {
 		try {
-			PostgresResource rsc = createPostgresServiceNull();
+			MysqlResource rsc = createMysqlServiceNull();
 			Response resp = rsc.delete();
 			assertThat(resp.getStatus(), is(500));
 		} catch (Exception e) {
