+++ application/wlp-web-java-spring/src/test/java/webapp/tier/util/BeforeAllTest.java
@@ -1,6 +1,6 @@
 package webapp.tier.util;
 
-import static org.junit.Assert.fail;
+import static org.junit.jupiter.api.Assertions.fail;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
