+++ application/wlp-web-java-spring/src/test/java/webapp/tier/util/GetConfigTest.java
@@ -1,7 +1,7 @@
 package webapp.tier.util;
 
-import static org.hamcrest.CoreMatchers.*;
-import static org.junit.Assert.*;
+import static org.hamcrest.CoreMatchers.is;
+import static org.hamcrest.MatcherAssert.assertThat;
 
 import org.junit.jupiter.api.Test;
